refactor(architecture): refactor likes
